requirements.txt
netmiko>=4.3.0
ntc-templates>=4.1.0
textfsm>=1.1.3
keyring>=24.0.0
pyyaml>=6.0.1
click>=8.1.7
